Resolve Project -> {Collaborator,Giver} asymmetry
The four criteria for our payment processor exemption to regulation as a
money transmitter are:[1]

  1. we must facilitate the purchase of goods or services, or the
     payment of bills for goods or services;

  2. we must operate through clearance and settlement systems that
     admit only BSA-regulated financial institutions;

  3. we must provide our service pursuant to a formal agreement; and

  4. our agreement must be at a minimum with the seller that
     provides the goods or services and receives the funds.

Therefore, I think it's a good idea to state right up front that we are
"a platform to facilitate payments for goods or services." This also
gives us a chance to define Work before referring to it later.

I don't believe that the "to facilitate this agreement" language that
this PR removes is entangled with our regulatory status. Instead, this
PR simplifies those clauses by referring to "any agreement."

[1] https://github.com/gratipay/inside.gratipay.com/issues/192#issuecomment-100048271